+++ src/arcIO.NET/Library.fs
@@ -1,5 +1,0 @@
-﻿namespace arcIO.NET
-
-module Say =
-    let hello name =
-        printfn "Hello %s" name